adds positional_observation_reporting, fixes lagrange parameters
